Board and BoardLogic were separated! idk how lucky i was
